La til flere Api funksjoner - Oppdatere location til User

+++ Hinder/src/main/java/Domain/Location.java
@@ -19,42 +19,44 @@ import lombok.NoArgsConstructor;
  *
  * @author Mikael
  */
-
-@Data @NoArgsConstructor
+@Data
+@NoArgsConstructor
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
 @Entity
 
 public class Location implements Serializable {
-    @Id @GeneratedValue
+
+    @Id
+    @GeneratedValue
     protected long id;
-    int latitude;
-    int longitude;
-    
-    
-    public Location(int latitude, int longitude){
-        this.latitude=latitude;
-        this.longitude=longitude;
-     
+    long latitude;
+    long longitude;
+
+    public Location(long latitude, long longitude) {
+        this.latitude = latitude;
+        this.longitude = longitude;
+
     }
-    public long getId(){
+
+    public long getId() {
         return id;
     }
 
-    public int getLatitude() {
+    public long getLatitude() {
         return latitude;
     }
 
-    public int getLongitude() {
+    public long getLongitude() {
         return longitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(long latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(long longitude) {
         this.longitude = longitude;
     }
-    
+
 }
